Knock off more warnings about signed/unsigned comparisons, plus a few shadowed declarations; also rename nSimplices_ to size_ in Isomorphism.

+++ engine/enumerate/hilbertdual.h
@@ -483,7 +483,7 @@ inline void HilbertDual::VecSpec<IntegerType, BitmaskType>::initNextHyp(
     nextHyp_ = 0;
 
     IntegerType tmp;
-    for (int i = 0; i < subspace.columns(); ++i)
+    for (size_t i = 0; i < subspace.columns(); ++i)
         if (subspace.entry(row, i) != 0 && (*this)[i] != 0) {
             tmp = subspace.entry(row, i);
             tmp *= (*this)[i];
